requirements.txt
huggingface_hub
python-dotenv
torch
torchvision
torchaudio
transformers
streamlit
asyncio
